+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - 29-12-2020
+
+* Added new method in Utils class for formatting number in given pattern.
+* Added documentation for showing how to use the widgets and utilities.
+
 ## [0.0.1] - 26-12-2020
 
 * Initial Release